Теперь 2 класса и починил отрицательные числа
